update jhdf5, factor out postprocessing hdf5, excuse missing native libs
