+++ pom.xml
@@ -109,9 +109,10 @@
 		<commons-discovery.version>0.5</commons-discovery.version>
 		<commons-io.version>2.4</commons-io.version>
 		<commons-lang.version>2.6</commons-lang.version>
-		<commons-lang3.version>3.12.0</commons-lang3.version>
+		<commons-lang3.version>3.13.0</commons-lang3.version>
 		<commons-math3.version>3.6.1</commons-math3.version>
 		<commons-pool2.version>2.5.0</commons-pool2.version>
+		<commons-text.version>1.11.0</commons-text.version>
 		<gson.version>2.9.1</gson.version>
 		<guava.version>33.0.0-jre</guava.version>
 		<guice.version>7.0.0</guice.version>
@@ -133,7 +134,7 @@
 		<jdom-contrib.version>1.1.3</jdom-contrib.version>
 		<jetty.version>9.4.29.v20200521</jetty.version>
 		<jfreechart.version>1.5.0</jfreechart.version>
-		<jhdf5.version>0.6.8</jhdf5.version>
+		<jhdf5.version>0.7.0-alpha</jhdf5.version>
 		<jhdf5-2-10.version>2.9</jhdf5-2-10.version>
 		<jmatio.version>1.0</jmatio.version>
 		<joda-time.version>2.3</joda-time.version>
