+++ vcell-core/src/main/java/cbit/vcell/simdata/SimulationData.java
@@ -981,8 +981,9 @@ public synchronized SimDataBlock getSimDataBlock(OutputContext outputContext, St
 			PDEDataInfo pdeDataInfo = new PDEDataInfo(vcDataId.getOwner(),vcDataId.getID(),varName,time,lastDataProcessingOutputInfoTime);
 			DataProcessingOutputDataValuesOP dataProcessingOutputDataValuesOP =
 				new DataProcessingOutputDataValuesOP(vcDataId, varName, TimePointHelper.createSingleTimeTimePointHelper(extractClosestPostProcessTime(time)), DataIndexHelper.createAllDataIndexesDataIndexHelper(), outputContext, null);
+			Hdf5DataProcessingReader hdf5DataProcessingReader = new Hdf5DataProcessingReader();
 			DataProcessingOutputDataValues dataProcessingOutputDataValues =
-				(DataProcessingOutputDataValues)DataSetControllerImpl.getDataProcessingOutput(dataProcessingOutputDataValuesOP, getDataProcessingOutputSourceFileHDF5());
+				(DataProcessingOutputDataValues) hdf5DataProcessingReader.getDataProcessingOutput(dataProcessingOutputDataValuesOP, getDataProcessingOutputSourceFileHDF5());
 			return new SimDataBlock(pdeDataInfo,dataProcessingOutputDataValues.getDataValues()[0]/*1 time only*/,VariableType.POSTPROCESSING);
 		}
 	}catch(Exception e){
@@ -1082,8 +1083,9 @@ synchronized double[][][] getSimDataTimeSeries0(
 				DataProcessingOutputDataValuesOP dataProcessingOutputDataValuesOP =
 					new DataProcessingOutputDataValuesOP(vcDataId, varNames[i],
 						TimePointHelper.createSpecificTimePointHelper(specificTimePoints), DataIndexHelper.createSpecificDataIndexHelper(indexes[i]), outputContext, null);
+				Hdf5DataProcessingReader hdf5DataProcessingReader = new Hdf5DataProcessingReader();
 				DataProcessingOutputDataValues dataProcessingOutputDataValues =
-					(DataProcessingOutputDataValues)DataSetControllerImpl.getDataProcessingOutput(dataProcessingOutputDataValuesOP, getDataProcessingOutputSourceFileHDF5());
+					(DataProcessingOutputDataValues) hdf5DataProcessingReader.getDataProcessingOutput(dataProcessingOutputDataValuesOP, getDataProcessingOutputSourceFileHDF5());
 				for (int j = 0; j < specificTimePoints.length; j++) {
 					results[j][i] = dataProcessingOutputDataValues.getDataValues()[j];
 				}
@@ -1809,7 +1811,8 @@ private void refreshDataProcessingOutputInfo(OutputContext outputContext) throws
 		lastOutputContext = outputContext;
 		DataProcessingOutputInfoOP dataProcessingOutputInfoOP = new DataProcessingOutputInfoOP(vcDataId, false, outputContext);
 		try{
-			dataProcessingOutputInfo = (DataProcessingOutputInfo)DataSetControllerImpl.getDataProcessingOutput(dataProcessingOutputInfoOP, dataProcessingOutputFile);
+			Hdf5DataProcessingReader hdf5DataProcessingReader = new Hdf5DataProcessingReader();
+			dataProcessingOutputInfo = (DataProcessingOutputInfo) hdf5DataProcessingReader.getDataProcessingOutput(dataProcessingOutputInfoOP, dataProcessingOutputFile);
 		}catch(Exception e){
 			throw new DataAccessException(e.getMessage(),(e.getCause()==null?e:e.getCause()));
 		}
